Setup Tranformation State, Animation Relay, started is hero form
Set up transformation (Temp hit L) animation will play and at the end the event will happen. Set up a animation relay that can be on a chile animator and pull any method that is in a script on the attached parent. may change if have a animation manager script.. for now works good may need to figure out a cching fix FUTURE. bool for is hero or not checks but need to set up so sprite changes when in hero form too.