General Clean up & Fixed Damage Delt
Unsure of all clean up... been a while... but everything works right now. (Player/enemy walking, attacking, death. Player transformed & all animations.) Damage fixed, it was dealing the damage value of their own (so player was taking the damage amount on the player when the enemy attacked them instead of taking the damage amount listed on the enemy to the player.