Fix enemy double attack, Start HP ui
Need to go back and make the UI Slider for health barm then continue setting up. Need to add a toggle or somethign so UI only appears when player has a skill unlocked to view the lower level enemy health bar. or a item, need to figure out... also is it a lot of memeory to have it and not display it or only add that ui once its been unlocked?